# chore: avoid unnecessary json marshal