@leoluz 2024-09-16
@agaudreault, @crenshaw-dev  The `TestRemarshalResources` seems to validate the type. However, it isn't clear to me what the outcome of `data` holding a different value could bring.