@crenshaw-dev 2024-09-17
@leoluz PTAL

The only way `data` is used is in unmarshaling to `reflect.New(reflect.TypeOf(item).Elem()).Interface()` (i.e. the schema type).

The "happy path" is already covered by `TestRemarshalResources`, i.e. we unmarshal into the schema type without an error. So the only uncovered behavior change is when there's some problem unmarshaling `data` into `unmarshalledObj`.

I've added tests to ensure the happy path still works when the input type isn't the `json` library's preferred type (i.e. `float` instead of `float64`).

I've also added a test to ensure that the unhappy path behaves the same (i.e. the input value can't be unmarshaled into the schema type).

I think this covers the relevant cases.