@crenshaw-dev 2024-09-17
Refactored the test to make SonarCloud happy.